cleaned up playground
