+++ playground/interactive_omim_RAG.py
@@ -1,106 +1,82 @@
-"""
-Chroma-based RAG model for gene-expert GPT on omim.org with chunking.
-
-USE WITHIN OPENAI LIMIT: String with maximum length 1048576
-"""
-
 import os
-import json
-from langchain.chains import ConversationalRetrievalChain
+import warnings
 from langchain_community.chat_models import ChatOpenAI
 from langchain_openai import OpenAIEmbeddings
 from langchain_community.vectorstores import Chroma
-# from langchain.memory import ConversationSummaryBufferMemory
-from langchain.schema import Document  # Import the Document class
+from langchain.schema import HumanMessage, SystemMessage
 
-import warnings
+warnings.filterwarnings("ignore")  # Suppress warnings
 
-warnings.filterwarnings("ignore")  # Suppress deprecation warnings
-
-os.environ["OPENAI_API_KEY"] = "YOUR KEY HERE"
-
-# Enable persistence to save the database to disk
-PERSIST = True
+os.environ["OPENAI_API_KEY"] = "YOUR API KEY HERE"
 
 # File paths
-data_path = "DATA JSON PATH"
-persist_directory_base = "DATABASE PATH"
-persist_directory = persist_directory_base  # Directory for the combined vector store
+persist_directory = "DATABASE DIRECTORY"  # Path to combined vector store
 
-# Step 1: Load chunked data from both sources
-print("Loading chunked data from both sources...")
-documents = []
-
-# Load gene data
-with open(data_path, "r", encoding="utf-8") as f:
-    for line in f:
-        entry = json.loads(line)
-        documents.append(entry)
-
-# Load cancer data
-# with open(cancer_json, "r", encoding="utf-8") as f:
-#     for line in f:
-#         entry = json.loads(line)
-#         documents.append(entry)
-
-print(f"Loaded {len(documents)} total chunks from omim database.")
-
-# Step 2: Initialize embeddings
+# Initialize embeddings
 embeddings = OpenAIEmbeddings()
 
-# Step 3: Create or load the combined vector store
-if PERSIST and os.path.exists(persist_directory):
+# Step 1: Load vector store
+if os.path.exists(persist_directory):
     print("Reusing existing combined database...")
     vectorstore = Chroma(persist_directory=persist_directory, embedding_function=embeddings)
 else:
-    print("Creating a new combined database...")
-    # Wrap each entry into a Document object
-    documents_wrapped = [
-        Document(page_content=doc['content'], metadata=doc['metadata']) for doc in documents
-    ]
-    vectorstore = Chroma.from_documents(
-        documents=documents_wrapped,  # Use the wrapped documents
-        embedding=embeddings,
-        persist_directory=persist_directory
-    )
-    if PERSIST:
-        vectorstore.persist()  # Save the combined database to disk
+    raise FileNotFoundError(f"Vector store not found at {persist_directory}. Ensure the data is preprocessed and saved.")
 
-# # Step 4: Set up the retrieval chain with memory
-# # memory buffer feature incompatible with conversation retrieval chain (see notes).
-# chat = ChatOpenAI(model="gpt-4o")
-# memory = ConversationSummaryBufferMemory(llm=chat, max_token_limit=5000)  # Summarization memory
-# retriever = vectorstore.as_retriever(search_kwargs={"k": 1})  # Retrieve top k most relevant answers
-# chain = ConversationalRetrievalChain.from_llm(
-#     llm=chat,
-#     retriever=retriever,
-#     memory=memory,
-# )
+# Step 2: Initialize retriever
+retriever = vectorstore.as_retriever(search_kwargs={"k": 5})  # Retrieve top 5 documents for context
 
-# # Step 5: Chat loop
-# print("Type 'quit', 'q', or 'exit' to end the chat.")
-# while True:
-#     query = input("Prompt: ")
-#     if query.lower() in ['quit', 'q', 'exit']:
-#         print("Goodbye!")
-#         break
-#
-#     # Run query through the chain
-#     result = chain({"question": query})
-#     print("Answer:", result['answer'])
+# Step 3: Initialize LLM
+llm = ChatOpenAI(model="gpt-4o")  # GPT model
 
+# Hybrid chain with fallback
+def hybrid_chain(query, retriever, llm, chat_history, max_length=4000):
+    """
+    Hybrid chain combining RAG with fallback to GPT general knowledge.
 
-# Step 4: Set up the retrieval chain without memory
-chat = ChatOpenAI(model="gpt-4o")
-retriever = vectorstore.as_retriever(search_kwargs={"k": 5})  # Retrieve top k most relevant answers
+    Parameters:
+    - query: User query.
+    - retriever: Retriever object for vector database.
+    - llm: GPT model.
+    - chat_history: List of previous interactions.
+    - max_length: Maximum character length for retrieved context.
 
-chain = ConversationalRetrievalChain.from_llm(
-    llm=chat,
-    retriever=retriever
-)
+    Returns:
+    - Answer string (text content only).
+    """
+    # Step 1: Retrieve relevant documents
+    retrieved_docs = retriever.get_relevant_documents(query)
 
-# Step 5: Chat loop using a simple chat_history list
-chat_history = []  # Initialize chat history as a list of (user, assistant) tuples
+    if retrieved_docs:
+        # Combine retrieved documents into context
+        context = "\n".join([f"{doc.metadata['gene_name']}: {doc.page_content}" for doc in retrieved_docs])
+        context = context[:max_length]  # Ensure the context is within LLM limits
+
+        # print(context)
+
+        # Create a prompt with retrieved context
+        messages = [
+            SystemMessage(content="You are an expert assistant with access to gene and cancer knowledge."),
+            HumanMessage(content=f"Using the following context, provide the most accurate and relevant answer to the question. "
+"Prioritize the provided context, but if the context does not contain enough information to fully address the question, "
+"use your best general knowledge to complete the answer:\n\n"
+            f"{context}\n\n"
+            f"Question: {query}")
+        ]
+        response = llm(messages)  # Pass structured messages
+        final_response = f"Document-Grounded Answer:\n{response.content}"
+    else:
+        # Fallback to GPT's general knowledge
+        messages = [
+            SystemMessage(content="You are an expert in cancer genomics and bioinformatics."),
+            HumanMessage(content=f"Answer the following question based on your general knowledge:\n\nQuestion: {query}")
+        ]
+        response = llm(messages)  # Pass structured messages
+        final_response = f"General Knowledge Answer:\n{response.content}"
+
+    return final_response
+
+# Step 4: Chat loop
+chat_history = []
 print("Type 'quit', 'q', or 'exit' to end the chat.")
 
 while True:
@@ -109,9 +85,9 @@ while True:
         print("Goodbye!")
         break
 
-    # Run query through the chain
-    result = chain({"question": query, "chat_history": chat_history})
-    print("Answer:", result['answer'])
+    # Get the hybrid response
+    answer = hybrid_chain(query, retriever, llm, chat_history)
+    print("Answer:", answer)
 
     # Update chat history
-    chat_history.append((query, result['answer']))
+    chat_history.append((query, answer))
